requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
openai==1.55.0
qdrant-client==1.17.0
pypdf==5.1.0
python-docx==1.1.2
google-auth==2.36.0
email-validator==2.3.0
requests==2.33.1
httpx==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
greenlet==3.1.1
aiofiles==24.1.0
